backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.11.*"
dependencies = [
    "accelerate>=1.13.0",
    "fastapi>=0.135.3",
    "onnxruntime>=1.23.1",
    "pillow>=12.2.0",
    "python-multipart>=0.0.22",
    "rembg>=2.0.67",
    "uvicorn[standard]>=0.42.0",
]
